LeftImage, RightImage components are functional
- References to CenteredImage, LeftImage, RightImage all updated across /blog, /docs, /projects pages
