+++ pages/blog/[year]/[id].js
@@ -11,7 +11,7 @@ import postStyles from '../../../styles/post.module.css';
  */
 import Link from 'next/link';
 import Image from 'next/image';
-import { CenteredImage, LeftImage } from '../../../components/content-helpers/CenteredImage';
+import { CenteredImage, LeftImage, RightImage } from '../../../components/content-helpers/images';
 import Button from '../../../components/button';
 import { docco, a11yDark } from 'react-syntax-highlighter/dist/cjs/styles/hljs'
 import SyntaxHighlighter from 'react-syntax-highlighter';
@@ -88,6 +88,7 @@ const Post = (props) => {
                     Link, 
                     CenteredImage, 
                     LeftImage,
+                    RightImage,
                     SyntaxHighlighter, 
                     MyHighlighter,
                     TikTok
