+++ posts/2022/pedernales-falls.mdx
@@ -33,13 +33,14 @@ what does this do?
 
 pretty water, but don't touch.
 
-<CenteredImage
+<RightImage
     srcUrl="https://meddlin-web.s3.us-east-2.amazonaws.com/2022-08-23_pedernales-falls/IMG_2763__q50.webp"
     width="4032"
     height="3024"
     alt="Wide-angle shot of blue-green colorful large puddle in dried riverbed."
-    caption="A single red chair...for reading."
-/>
+    caption="A single red chair...for reading.">
+    what about this?
+</RightImage>
 
 water with rock in hole?
